requirements.txt
alpaca-py
pandas
pytz
datetime
dotenv
argparse
numpy
